httpd: Add root handlers for `/api/v1` and `/api`
